feat(secrets): add Azure provider + WIF config
