--- cloud/gcp/tofu/k3s-wif/keys.json
@@ -1,1 +1,0 @@
-{"keys":[{"use":"sig","kty":"RSA","kid":"4UrU-PfXX9-ZBNKFj5nOZ1u3hPnXJE1UkwFen3RhfK4","alg":"RS256","n":"xUXOZtY3q4itPbVyYL_u7rslZ-fULZ3RBmpXT4wTgK1sImc8GxAF2-XVklFvOpnqFlrYaXC40GMULbBdM8zCdoGWtky6bSuFwojlY0xjs6lW_YbFiRDWTngwpwHCY18WtZngcP-8kQynfzqF_S8zw_uVg1H6Ko-QzNVnef9iocim5v3AizHJvAV0d9WwhGGDQwuDPxp9C__6slX25ZmaFTi9khO5yNia_lKn4SA67IK8ccs7RDlrUpGqxBK6DiiKXvOLjeHodaYE-XqN_anqx0GOtCVtchGJor7Y0A9X0RQBJcexOk13grmZpsbJ8zFXTgPPxIOXMDgFRwsXVO8zGw","e":"AQAB"}]}
