+++ cloud/tofu/backend.tf
@@ -8,6 +8,14 @@ terraform {
       source  = "hashicorp/archive"
       version = "2.6.0"
     }
+    azuread = {
+      source = "hashicorp/azuread"
+      version = "3.0.2"
+    }
+    azurerm = {
+      source = "hashicorp/azurerm"
+      version = "4.4.0"
+    }
     google = {
       source  = "hashicorp/google"
       version = "6.6.0"
