+++ cloud/tofu/k3s-wif/.well-known/openid-configuration
