+++ kubernetes/workloads/secrets-store-csi-driver-provider-azure/manifests/base/kustomization.yaml
@@ -1,0 +1,7 @@
+---
+kind: Kustomization
+apiVersion: kustomize.config.k8s.io/v1beta1
+resources:
+  - daemonset.yaml
+  - namespace.yaml
+  - serviceaccount.yaml
