+++ kubernetes/workloads/secrets-store-csi-driver-provider-azure/manifests/base/namespace.yaml
@@ -1,0 +1,7 @@
+---
+# yamllint disable rule:line-length
+---
+apiVersion: v1
+kind: Namespace
+metadata:
+  name: secrets-store-csi-driver
