+++ kubernetes/workloads/secrets-store-csi-driver-provider-azure/manifests/base/serviceaccount.yaml
@@ -1,0 +1,14 @@
+---
+# yamllint disable rule:line-length
+apiVersion: v1
+kind: ServiceAccount
+metadata:
+  name: csi-secrets-store-provider-azure
+  namespace: secrets-store-csi-driver
+  labels:
+    app.kubernetes.io/instance: csi-secrets-store-provider-azure
+    app.kubernetes.io/managed-by: Helm
+    app.kubernetes.io/version: 1.5.3
+    helm.sh/chart: csi-secrets-store-provider-azure-1.5.6
+    app.kubernetes.io/name: csi-secrets-store-provider-azure
+    app: csi-secrets-store-provider-azure
