--- kubernetes/workloads/secrets-store-csi-driver-provider-gcp/manifests/pod-adc/secretproviderclass.yaml
@@ -1,12 +1,0 @@
-apiVersion: secrets-store.csi.x-k8s.io/v1
-kind: SecretProviderClass
-metadata:
-  name: app-secrets
-  namespace: default
-spec:
-  provider: gcp
-  parameters:
-    auth: pod-adc
-    secrets: |
-      - resourceName: "projects/ben-homelab/secrets/test-secret/versions/latest"
-        path: "good1.txt"
