+++ kubernetes/workloads/secrets-store-csi-driver/testing/secretproviderclass.yaml
@@ -1,0 +1,32 @@
+apiVersion: secrets-store.csi.x-k8s.io/v1
+kind: SecretProviderClass
+metadata:
+  name: gcp
+  namespace: default
+spec:
+  provider: gcp
+  parameters:
+    auth: pod-adc
+    secrets: |
+      - resourceName: "projects/ben-homelab/secrets/test-secret/versions/latest"
+        path: "good1.txt"
+---
+apiVersion: secrets-store.csi.x-k8s.io/v1
+kind: SecretProviderClass
+metadata:
+  name: azure
+  namespace: default
+spec:
+  provider: azure
+  parameters:
+    usePodIdentity: "false"
+    useVMManagedIdentity: "false"
+    tenantID: "d56a3b3a-f480-4f7c-ae4b-19a0784de9eb"
+    clientID: "a462129c-4106-450a-8fa5-40012c2cc9cd"
+    keyvaultName: "dronenb-homelab"
+    objects: |
+      array:
+        - |
+          objectName: secret-sauce
+          objectType: secret
+          objectVersion: ""
